// mobile/types.tsx
export type RootStackParamList = {
  Root: NavigatorScreenParams<RootTabParamList> | undefined;
  AddWorkOrder: { location?: Location; asset?: AssetDTO };
  WorkOrderStats: undefined;
  EditWorkOrder: { workOrder: WorkOrder; tasks: Task[] };
  EditRequest: { request: Request };
  EditAsset: { asset: AssetDTO };
  EditLocation: { location: Location };
  EditPart: { part: Part };
  EditMeter: { meter: Meter };
  EditUser: { user: User };
  EditCustomer: { customer: Customer };
  EditVendor: { vendor: Vendor };
  EditTeam: { team: Team };
  AddRequest: undefined;
  AddAsset: {
    location?: Location;
    parentAsset?: AssetDTO;
    nfcId?: string;
    barCode?: string;
  };
  AddLocation: undefined;
  AddPart: undefined;
  AddMeter: undefined;
  AddUser: undefined;
  WODetails: { id: number; workOrderProp?: WorkOrder };
  AssetDetails: { id: number; assetProp?: AssetDTO };
  LocationDetails: { id: number; locationProp?: Location };
  RequestDetails: { id: number; requestProp?: Request };
  UserDetails: { id: number; userProp?: OwnUser };
  UserProfile: undefined;
  TeamDetails: { id: number; teamProp?: Team };
  PartDetails: { id: number; partProp?: Part };
  MeterDetails: { id: number; meterProp?: Meter };
  CustomerDetails: { id: number; customerProp?: Customer };
  VendorDetails: { id: number; vendorProp?: Vendor };
  Modal: undefined;
  Tasks: {
    tasksProps: Task[];
    workOrderId: number;
  };
  SelectParts: { onChange: (parts: PartMiniDTO[]) => void; selected: number[] };
  SelectCustomers: {
    onChange: (customers: CustomerMiniDTO[]) => void;
    selected: number[];
    multiple: boolean;
  };
  SelectVendors: {
    onChange: (vendors: VendorMiniDTO[]) => void;
    selected: number[];
    multiple: boolean;
  };
  SelectUsers: {
    onChange: (users: UserMiniDTO[]) => void;
    selected: number[];
    multiple: boolean;
  };
  SelectTeams: {
    onChange: (teams: TeamMiniDTO[]) => void;
    selected: number[];
    multiple: boolean;
  };
  SelectMeters: {
    onChange: (meters: MeterMiniDTO[]) => void;
    selected: number[];
    multiple: boolean;
  };
  SelectLocations: {
    onChange: (locations: LocationMiniDTO[]) => void;
    selected: number[];
    multiple: boolean;
  };
  SelectAssets: {
    onChange: (assets: AssetMiniDTO[]) => void;
    selected: number[];
    multiple: boolean;
    locationId: number | null;
  };
  SelectTasks: { onChange: (tasks: Task[]) => void; selected: Task[] };
  SelectChecklists: { onChange: (tasks: Task[]) => void; selected: Task[] };
  SelectTasksOrChecklist: {
    onChange: (tasks: Task[]) => void;
    selected: Task[];
  };
  SelectCategories: {
    onChange: (categories: Category[]) => void;
    selected: number[];
    multiple: boolean;
    type: string;
  };
  SelectNfc: { onChange: (value: string) => void };
  SelectBarcode: { onChange: (value: string) => void };
  CompleteWorkOrder: {
    onComplete: (
      signature: string | undefined,
      feedback: string | undefined
    ) => Promise<any>;
    fieldsConfig: { feedback: boolean; signature: boolean };
  };
  NotFound: undefined;
  Meters: undefined;
  Parts: undefined;
  Locations: { id?: number; hierarchy?: number[] };
  Assets: {
    id?: number;
    hierarchy?: number[];
    filterFields?: FilterField[];
    view?: 'list' | 'hierarchy';
  };
  PeopleTeams: undefined;
  VendorsCustomers: undefined;
  Notifications: undefined;
  Settings: undefined;
  WorkOrderFilters: {
    filterFields: FilterField[];
    onFilterChange: (filterFields: FilterField[]) => void;
  };
  AddAdditionalCost: { workOrderId: number };
  AddAdditionalTime: { workOrderId: number };
  ScanAsset: undefined;
};

export type RootStackScreenProps<Screen extends keyof RootStackParamList> =
  NativeStackScreenProps<RootStackParamList, Screen>;

export type RootTabParamList = {
  Home: undefined;
  WorkOrders: { filterFields: FilterField[]; fromHome?: boolean };
  AddEntities: undefined;
  Requests: undefined;
  MoreEntities: undefined;
};